refactor: Update TabPanels className in Tabs component
